+++ src/main/kotlin/dev/adirelle/adicrafter/crafter/api/power/PowerGenerator.kt
@@ -13,4 +13,6 @@ interface PowerGenerator : SingleViewStorage<PowerVariant> {
     fun isActive(): Boolean = true
     fun tick(world: World): Boolean = false
     fun setUpdateCallback(callback: () -> Unit) {}
+    
+    fun hasPowerBar(): Boolean
 }