+++ src/main/kotlin/dev/adirelle/adicrafter/crafter/impl/CrafterDataAccessorAdapter.kt
@@ -17,7 +17,8 @@ class CrafterDataAccessorAdapter(
         const val GRID_LAST_SLOT = GRID_FIRST_SLOT + Grid.SIZE - 1
         const val FORECAST_SLOT = GRID_LAST_SLOT + 1
         const val RESULT_SLOT = FORECAST_SLOT + 1
-        const val INVENTORY_SIZE = RESULT_SLOT + 1
+        const val GENERATOR_SLOT = RESULT_SLOT + 1
+        const val INVENTORY_SIZE = GENERATOR_SLOT + 1
 
         const val RECIPE_FLAGS_PROP = 0
         const val MISSING_PROP = 1
@@ -80,25 +81,31 @@ class CrafterDataAccessorAdapter(
             inventories.forEach { it.onClose(player) }
         }
 
-        private inline fun <T> withInventory(index: Int, block: Inventory.(Int) -> T): T =
+        private inline fun <T> withInventory(index: Int, default: T, block: Inventory.(Int) -> T): T =
             when (index) {
                 in GRID_FIRST_SLOT..GRID_LAST_SLOT -> grid.block(index - GRID_FIRST_SLOT)
                 FORECAST_SLOT                      -> forecast.block(0)
                 RESULT_SLOT                        -> result.block(0)
+                GENERATOR_SLOT                     -> generator?.block(0) ?: default
                 else                               -> throw IndexOutOfBoundsException()
             }
 
+        private inline fun withInventory(index: Int, block: Inventory.(Int) -> Unit) {
+            withInventory(index, Unit, block)
+        }
+
         override fun getStack(slot: Int): ItemStack =
-            withInventory(slot) { getStack(it) }
+            withInventory(slot, ItemStack.EMPTY) { getStack(it) }
 
         override fun removeStack(slot: Int, amount: Int) =
-            withInventory(slot) { removeStack(it, amount) }
+            withInventory(slot, ItemStack.EMPTY) { removeStack(it, amount) }
 
         override fun removeStack(slot: Int): ItemStack =
-            withInventory(slot) { removeStack(it) }
+            withInventory(slot, ItemStack.EMPTY) { removeStack(it) }
 
-        override fun setStack(slot: Int, stack: ItemStack) =
+        override fun setStack(slot: Int, stack: ItemStack) {
             withInventory(slot) { setStack(it, stack) }
+        }
 
     }